Fix ODG

+++ asivamosffie.model/Models/VContratoPagosRealizados.cs
@@ -8,11 +8,11 @@ namespace asivamosffie.model.Models
         public int ContratoId { get; set; }
         public bool? EsPreconstruccion { get; set; }
         public int ValorFacturado { get; set; }
-        public int? PorcentajeFacturado { get; set; }
+        public decimal? PorcentajeFacturado { get; set; }
         public string FaseContrato { get; set; }
         public int? PagosRealizados { get; set; }
         public decimal? SaldoPorPagar { get; set; }
-        public int? PorcentajePorPagar { get; set; }
+        public decimal? PorcentajePorPagar { get; set; }
         public decimal? ValorSolicitud { get; set; }
     }
 }
